+++ src/editors/array-table-object.js
@@ -28,12 +28,12 @@ class EditorArrayTableObject extends EditorArray {
   }
 
   addEventListeners () {
-    super.addEventListeners()
-
     this.control.addBtn.addEventListener('click', () => {
       this.activeItemIndex = this.instance.value.length
       this.instance.addItem('user')
     })
+
+    this.addJsonDataEventListeners()
   }
 
   isSortable () {