+++ src/editors/array-table.js
@@ -12,11 +12,12 @@ class EditorArrayTable extends EditorArray {
   }
 
   addEventListeners () {
-    super.addEventListeners()
     this.control.addBtn.addEventListener('click', () => {
       this.activeItemIndex = this.instance.value.length
       this.instance.addItem('user')
     })
+
+    this.addJsonDataEventListeners()
   }
 
   isSortable () {